feat: improve storage paths with human-readable names and fix course announcement filtering
- Update storage system to use course/lesson/module/assignment names instead of IDs in directory paths
- Add sanitizeDirectoryName helper for URL-safe directory names
- Update uploadMediaFile to fetch and use names from database for all entity types
- Fix course announcement filtering to exclude global/general announcements
- Update StudentLessonViewer and CourseDetailsPage to use consistent filtering logic
- Require contextType !== 'general' in addition to audienceScope === 'course' for course-specific views
- Update announcement system documentation with new filtering requirements
- Ensure filename sanitization is applied to all uploads
- Store entity names in media metadata for reference

+++ database/schema.sql
@@ -330,7 +330,7 @@ CREATE TABLE IF NOT EXISTS media_files (
     filename TEXT NOT NULL,
     original_filename TEXT NOT NULL,
     file_type TEXT NOT NULL,
-    file_size INTEGER NOT NULL,
+    file_size BIGINT NOT NULL,
     mime_type TEXT NOT NULL,
     file_url TEXT NOT NULL,
     thumbnail_url TEXT,
